initial work on agent
