+++ agent/urls.py
@@ -13,10 +13,6 @@
 # limitations under the License.
 from django.conf.urls import patterns, url
 
-urlpatterns = patterns('hosts.views',
-    url(r'^$', 'index'),
-    url(r'^edit/(?P<host_id>.*)/$', 'edit_host'),
-    url(r'^enable/(?P<host_id>.*)/$', 'enable_host'),
-    url(r'^disable/(?P<host_id>.*)/$', 'disable_host'),
-    url(r'^remove/(?P<host_id>.*)/$', 'remove_host'),
+urlpatterns = patterns('agent.views',
+    url(r'^register/$', 'register', name='agent.register'),
 )
