+++ hosts/templates/hosts/edit_host.html
@@ -6,7 +6,7 @@
 {% block content %}
 <div class="panel panel-default">
     <div class="panel-heading heading-large">
-        <span class="panel-title section-title">{% trans 'Add Host' %}</span>
+        <span class="panel-title section-title">{% trans 'Edit Host' %}</span>
         <span class="pull-right">
         </span>
     </div>
