+++ shipyard/urls.py
@@ -29,6 +29,7 @@ urlpatterns = patterns('',
     url(r'^$', 'shipyard.views.index', name='index'),
     url(r'^api/login', 'accounts.views.api_login', name='api_login'),
     url(r'^api/', include(v1_api.urls)),
+    url(r'^agent/', include('agent.urls')),
     url(r'^accounts/', include('accounts.urls')),
     url(r'^applications/', include('applications.urls')),
     url(r'^containers/', include('containers.urls')),
